requirements.txt
streamlit
langchain
faiss-cpu
tiktoken
PyPDF2
langchain-google-genai
python-dotenv
langchain-community
sentence-transformers